refactor(core): fold sports-api and predictions-api into core-api [REL-15]
Second and third folds per ADR-0002, using the REL-14 localSource mold.
Route shapes, cache keys, TTLs, and response bodies are identical; the
desktop app needs no changes. Both Rust ingesters are untouched.

- api/core/sports.go: /sports routes (auth user now comes from LogtoAuth
  via GetUserID, never the X-User-Sub header the proxy used to inject),
  fair-share dashboard allocator, league catalog/standings/teams,
  health probe via INTERNAL_SPORTS_URL
- api/core/predictions.go: /predictions routes incl. the candlesticks
  pass-through to the Rust service (5-min cache), shared-cache dashboard
  default path, narrowed per-user view for pre-v1.1.5 configs
- Reused core's existing extractLeaguesFromConfig; containsMonth became
  slices.Contains per the audit backlog
- Dropped per REL-10: both /internal/cdc handlers and all subscriber-set
  writes (sports league sets keep their core-side writers until REL-17)
- Removed: channels/{sports,predictions}/api modules, both k8s manifests,
  NetworkPolicy pairs (both Rust services now admit core-api), deploy
  matrix rows + path filters, backend-tests entries, smoke entries,
  compose blocks; INTERNAL_{SPORTS,PREDICTIONS}_URL moved to core-config
- Docs: AGENTS.md, LOCAL_SETUP.md

Post-merge manual cleanup (kubectl apply does not prune):
  kubectl delete deployment/sports-api service/sports-api deployment/predictions-api service/predictions-api -n scrollr
  kubectl delete networkpolicy allow-core-to-sports-api allow-sports-api-to-sports-service allow-core-to-predictions-api allow-predictions-api-to-predictions-service -n scrollr
