+++ LOCAL_SETUP.md
@@ -32,8 +32,9 @@ windows. `make help` lists everything. Day-to-day:
 | `make clean` | Stop **and wipe** the Postgres + Redis volumes. |
 | `make web` / `make desktop` | Run the front-ends natively. |
 
-**Ports** (host): core `18080` (serves `/finance/*` natively per ADR-0002), rss `8083`, sports `8082`,
-fantasy `8084`, predictions `8085`, site `3000`. Postgres `5432`, Redis `6379`.
+**Ports** (host): core `18080` (serves `/finance/*`, `/sports/*`, and
+`/predictions/*` natively per ADR-0002), rss `8083`, fantasy `8084`,
+site `3000`. Postgres `5432`, Redis `6379`.
 
 ### Prereqs (one-time)
 
@@ -146,11 +147,12 @@ against the **local** API with your real account + tier.
 
 ## Milestone 2 — Live channel data (all four channels local)
 
-All four data channels run locally against the **same** Postgres/Redis as the
-Core API. sports/rss/fantasy each have a **Go API** (registers itself in
-Redis; the Core discovers + proxies it); finance's Go API was folded into
-core (ADR-0002). finance/sports/rss each have a **Rust ingester** (polls the
-provider, writes rows). Fantasy is Go-native (in-process sync).
+All data channels run locally against the **same** Postgres/Redis as the
+Core API. rss and fantasy each have a **Go API** (registers itself in
+Redis; the Core discovers + proxies it); the finance, sports, and
+predictions Go APIs were folded into core (ADR-0002). finance/sports/rss/
+predictions each have a **Rust ingester** (polls the provider, writes
+rows). Fantasy is Go-native (in-process sync).
 
 ### Ports & secrets
 
@@ -158,8 +160,9 @@ provider, writes rows). Fantasy is Go-native (in-process sync).
 |---|---|---|---|
 | rss     | 8083     | 3004 | none (public feeds) |
 | finance | (in core) | 3001 | `TWELVEDATA_API_KEY` |
-| sports  | 8082     | 3002 | `API_SPORTS_KEY` |
+| sports  | (in core) | 3002 | `API_SPORTS_KEY` |
 | fantasy | 8084     | —    | `YAHOO_CLIENT_ID` / `YAHOO_CLIENT_SECRET` |
+| predictions | (in core) | 3005 | Kalshi key via `make prep` |
 
 Provider keys live in the DO cluster secret `scrollr-secrets`
 (namespace `scrollr`); pull them into gitignored `channels/<ch>/.env` with
@@ -183,21 +186,21 @@ processes read it and would collide; pass the Go API's port inline instead.
 ```bash
 GO="$LOCALAPPDATA/go-toolchain/go/bin/go.exe"
 
-# rss  (no key)          finance (Rust only — API lives in core, ADR-0002)   sports    fantasy (Go-native)
+# rss (no key — the only remaining channel Go API besides fantasy)   fantasy (Go-native)
 ( cd channels/rss/service     && set -a && . ../.env && set +a && cargo run )              # :3004
 ( cd channels/rss/api         && set -a && . ../.env && set +a && "$GO" run . )            # :8083
 ( cd channels/finance/service && set -a && . ../.env && set +a && cargo run )              # :3001
 ( cd channels/sports/service  && set -a && . ../.env && set +a && cargo run )              # :3002
-( cd channels/sports/api      && set -a && . ../.env && set +a && "$GO" run . )            # :8082
 ( cd channels/fantasy/api     && set -a && . ../.env && set +a && "$GO" run . )            # :8084
 ```
 
-Finance's Go API was folded into core (ADR-0002, REL-14): core serves
-`/finance/*` itself and probes the Rust ingester via
-`INTERNAL_FINANCE_URL` (set it in `api/.env`, e.g.
-`http://localhost:3001`, or core reports finance healthy without probing).
+The finance, sports, and predictions Go APIs were folded into core
+(ADR-0002): core serves `/finance/*`, `/sports/*`, and `/predictions/*`
+itself and probes the Rust ingesters via `INTERNAL_{FINANCE,SPORTS,
+PREDICTIONS}_URL` (set them in `api/.env`, e.g. `http://localhost:3001`;
+unset means core reports that source healthy without probing).
 
-The Core logs `[Discovery] Channels updated: 3 active [fantasy,rss,sports]`
+The Core logs `[Discovery] Channels updated: 2 active [fantasy,rss]`
 within ~10s of each Go API starting.
 
 ### Gotchas (hit and fixed)
