--- channels/sports/docker-compose.dev.yml
@@ -1,27 +1,0 @@
-version: "3.8"
-
-services:
-  scrollr-sports-api:
-    build:
-      context: ./api
-      dockerfile: Dockerfile
-    environment:
-      - DATABASE_URL=${DATABASE_URL}
-      - REDIS_URL=${REDIS_URL}
-      - ALLOWED_ORIGINS=${ALLOWED_ORIGINS}
-      - CHANNEL_URL=https://sports.api.myscrollr.enanimate.dev
-      - INTERNAL_SPORTS_URL=http://scrollr-sports-service:3002
-    depends_on:
-      scrollr-sports-service:
-        condition: service_started
-    restart: unless-stopped
-
-  scrollr-sports-service:
-    build:
-      context: ./service
-      dockerfile: Dockerfile
-    environment:
-      - DATABASE_URL=${DATABASE_URL}
-      - API_SPORTS_KEY=${API_SPORTS_KEY:-test_api_sports_key}
-      - API_SPORTS_BASE_URL=http://mock-apisports:9004
-    restart: unless-stopped
