--- channels/sports/docker-compose.yml
@@ -1,35 +1,0 @@
-version: "3.8"
-
-services:
-  scrollr-sports-api:
-    build:
-      context: ./api
-      dockerfile: Dockerfile
-    container_name: scrollr-sports-api
-    # Bind only to localhost — the core gateway is the only trusted caller
-    # and reaches us via the Docker network (CHANNEL_URL). Publishing on
-    # 0.0.0.0 let anything else on the host bypass the gateway and hit
-    # our routes directly (channels don't validate JWTs — they trust the
-    # X-User-Sub header set by the gateway).
-    ports:
-      - "127.0.0.1:8082:8082"
-    environment:
-      - DATABASE_URL=${DATABASE_URL}
-      - REDIS_URL=${REDIS_URL}
-      - ENCRYPTION_KEY=${ENCRYPTION_KEY}
-      - ALLOWED_ORIGINS=${ALLOWED_ORIGINS}
-      - CHANNEL_URL=${CHANNEL_URL}
-      - INTERNAL_SPORTS_URL=http://scrollr-sports-service:3002
-    restart: unless-stopped
-
-  scrollr-sports-service:
-    build:
-      context: ./service
-      dockerfile: Dockerfile
-    container_name: scrollr-sports-service
-    ports:
-      - "3002:3002"
-    environment:
-      - DATABASE_URL=${DATABASE_URL}
-      - API_SPORTS_KEY=${API_SPORTS_KEY}
-    restart: unless-stopped
