--- k8s/predictions-api.yaml
@@ -1,126 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: predictions-api
-  namespace: scrollr
-  labels:
-    app: predictions-api
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: predictions-api
-  template:
-    metadata:
-      labels:
-        app: predictions-api
-    spec:
-      containers:
-        - name: predictions-api
-          image: registry.digitalocean.com/scrollr/predictions-api:latest
-          ports:
-            - containerPort: 8085
-          env:
-            - name: CHANNEL_URL
-              valueFrom:
-                configMapKeyRef:
-                  name: channels-config
-                  key: PREDICTIONS_CHANNEL_URL
-            - name: INTERNAL_PREDICTIONS_URL
-              valueFrom:
-                configMapKeyRef:
-                  name: channels-config
-                  key: INTERNAL_PREDICTIONS_URL
-            - name: ALLOWED_ORIGINS
-              valueFrom:
-                configMapKeyRef:
-                  name: channels-config
-                  key: ALLOWED_ORIGINS
-            - name: DATABASE_URL
-              valueFrom:
-                secretKeyRef:
-                  name: scrollr-secrets
-                  key: DATABASE_URL
-            - name: REDIS_URL
-              valueFrom:
-                secretKeyRef:
-                  name: scrollr-secrets
-                  key: REDIS_URL
-            - name: PORT
-              value: "8085"
-            # Sentry — set PREDICTIONS_API_SENTRY_DSN in scrollr-secrets once
-            # a Sentry project exists (other APIs inline per-service DSNs).
-            - name: SENTRY_DSN
-              valueFrom:
-                secretKeyRef:
-                  name: scrollr-secrets
-                  key: PREDICTIONS_API_SENTRY_DSN
-                  optional: true
-            - name: SENTRY_USER_SALT
-              valueFrom:
-                secretKeyRef:
-                  name: scrollr-secrets
-                  key: SENTRY_USER_SALT
-                  optional: true
-            - name: ENVIRONMENT
-              valueFrom:
-                secretKeyRef:
-                  name: scrollr-secrets
-                  key: ENVIRONMENT
-                  optional: true
-            - name: GIT_SHA
-              valueFrom:
-                secretKeyRef:
-                  name: scrollr-secrets
-                  key: GIT_SHA
-                  optional: true
-          resources:
-            requests:
-              cpu: 50m
-              memory: 64Mi
-            limits:
-              cpu: 250m
-              memory: 128Mi
-          # /internal/health pings DB, Redis, and the Rust ingestion
-          # service's /health/ready. Returns 503 when any dependency is down.
-          startupProbe:
-            httpGet:
-              path: /internal/health
-              port: 8085
-            initialDelaySeconds: 3
-            periodSeconds: 5
-            failureThreshold: 24
-            timeoutSeconds: 3
-          livenessProbe:
-            httpGet:
-              path: /internal/health
-              port: 8085
-            initialDelaySeconds: 10
-            periodSeconds: 30
-            failureThreshold: 6
-            timeoutSeconds: 3
-          readinessProbe:
-            httpGet:
-              path: /internal/health
-              port: 8085
-            initialDelaySeconds: 5
-            periodSeconds: 10
-            failureThreshold: 3
-            timeoutSeconds: 3
-          lifecycle:
-            preStop:
-              exec:
-                command: ["/bin/sh", "-c", "sleep 10"]
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: predictions-api
-  namespace: scrollr
-spec:
-  selector:
-    app: predictions-api
-  ports:
-    - port: 8085
-      targetPort: 8085
-  type: ClusterIP
