--- k8s/sports-api.yaml
@@ -1,127 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: sports-api
-  namespace: scrollr
-  labels:
-    app: sports-api
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: sports-api
-  template:
-    metadata:
-      labels:
-        app: sports-api
-    spec:
-      containers:
-        - name: sports-api
-          image: registry.digitalocean.com/scrollr/sports-api:latest
-          ports:
-            - containerPort: 8082
-          env:
-            - name: CHANNEL_URL
-              valueFrom:
-                configMapKeyRef:
-                  name: channels-config
-                  key: SPORTS_CHANNEL_URL
-            - name: INTERNAL_SPORTS_URL
-              valueFrom:
-                configMapKeyRef:
-                  name: channels-config
-                  key: INTERNAL_SPORTS_URL
-            - name: ALLOWED_ORIGINS
-              valueFrom:
-                configMapKeyRef:
-                  name: channels-config
-                  key: ALLOWED_ORIGINS
-            - name: DATABASE_URL
-              valueFrom:
-                secretKeyRef:
-                  name: scrollr-secrets
-                  key: DATABASE_URL
-            - name: REDIS_URL
-              valueFrom:
-                secretKeyRef:
-                  name: scrollr-secrets
-                  key: REDIS_URL
-            - name: ENCRYPTION_KEY
-              valueFrom:
-                secretKeyRef:
-                  name: scrollr-secrets
-                  key: ENCRYPTION_KEY
-            # Sentry — per-service DSN inlined (ingestion-only, not a leak).
-            - name: SENTRY_DSN
-              value: "https://1fe6d97731f2f049b6bed79af12553cd@o4511384091033600.ingest.us.sentry.io/4511384397086720"
-            # Shared Sentry config from scrollr-secrets (optional=true so
-            # the Deployment doesn't fail if the keys aren't populated yet).
-            - name: SENTRY_USER_SALT
-              valueFrom:
-                secretKeyRef:
-                  name: scrollr-secrets
-                  key: SENTRY_USER_SALT
-                  optional: true
-            - name: ENVIRONMENT
-              valueFrom:
-                secretKeyRef:
-                  name: scrollr-secrets
-                  key: ENVIRONMENT
-                  optional: true
-            - name: GIT_SHA
-              valueFrom:
-                secretKeyRef:
-                  name: scrollr-secrets
-                  key: GIT_SHA
-                  optional: true
-          resources:
-            requests:
-              cpu: 50m
-              memory: 64Mi
-            limits:
-              cpu: 250m
-              memory: 128Mi
-          # /internal/health now pings DB, Redis, and the Rust ingestion
-          # service's /health/ready (see PR #108). It returns 503 when
-          # any dependency is down.
-          startupProbe:
-            httpGet:
-              path: /internal/health
-              port: 8082
-            initialDelaySeconds: 3
-            periodSeconds: 5
-            failureThreshold: 30
-            timeoutSeconds: 3
-          livenessProbe:
-            httpGet:
-              path: /internal/health
-              port: 8082
-            initialDelaySeconds: 10
-            periodSeconds: 30
-            failureThreshold: 6
-            timeoutSeconds: 3
-          readinessProbe:
-            httpGet:
-              path: /internal/health
-              port: 8082
-            initialDelaySeconds: 5
-            periodSeconds: 10
-            failureThreshold: 3
-            timeoutSeconds: 3
-          lifecycle:
-            preStop:
-              exec:
-                command: ["/bin/sh", "-c", "sleep 10"]
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: sports-api
-  namespace: scrollr
-spec:
-  selector:
-    app: sports-api
-  ports:
-    - port: 8082
-      targetPort: 8082
-  type: ClusterIP
